+++ backend/substrapp/urls.py
@@ -40,6 +40,9 @@ compute_plan_router.register(
 compute_plan_router.register(r"testtuple", views.CPTaskViewSet, basename=f"{CP_BASENAME_PREFIX}testtuple")
 compute_plan_router.register(r"algos", views.CPAlgoViewSet, basename=f"{CP_BASENAME_PREFIX}algo")
 compute_plan_router.register(r"perf", views.CPPerformanceViewSet, basename=f"{CP_BASENAME_PREFIX}perf")
+compute_plan_router.register(
+    r"workflow_graph", views.CPWorkflowGraphViewSet, basename=f"{CP_BASENAME_PREFIX}workflow_graph"
+)
 
 
 urlpatterns = [